+++ Projet/CrystalGate/CrystalGate/Objet.cs
@@ -48,7 +48,7 @@ namespace CrystalGate
         public float LastAttack;
         public Unite OlduniteAttacked;
         public Unite uniteAttacked;
-        public byte id;
+        public int id;
 
         public Objet(Vector2 Position)
         {
