+++ Projet/CrystalGate/CrystalGate/Reseau/Player.cs
@@ -12,14 +12,14 @@ namespace CrystalGate
         public bool Mooved;
         public float objectifPointX; // Le point ou cherche à aller le joueur
         public float objectifPointY;
-        public byte idUniteAttacked; // L'unité attaqué par le joueur
+        public int idUniteAttacked; // L'unité attaqué par le joueur
 
         public byte idSortCast; // Le sort cast par le joueur
         public float pointSortX;
         public float pointSortY;
-        public byte idUniteCibleCast;
+        public int idUniteCibleCast;
 
-        public byte[] LastDeath = new byte[10];
+        public int[] LastDeath = new int[10];
         public int LastItemUsed = -1;
         public int LastStuffUsed = -1;
         public string level;
